continue to refactor gen_peqs
