+++ generate_peqs.py
@@ -55,324 +55,19 @@ import flammkuchen as fl
 import numpy as np
 import pandas as pd
 import ray
-import scipy.optimize as opt
 
 from generate_common import get_custom_logger, args2level, custom_ray_init
 from datas.metadata import speakers_info as metadata
 from spinorama.ltype import Vector, Peq
-from spinorama.filter_iir import Biquad
-from spinorama.filter_peq import peq_build  # peq_print
 from spinorama.load_rewseq import parse_eq_iir_rews
 from spinorama.filter_peq import peq_format_apo
 from spinorama.filter_scores import scores_apply_filter, scores_print2
-from spinorama.auto_loss import loss, score_loss
 from spinorama.auto_target import get_freq, get_target
-from spinorama.auto_range import (
-    propose_range_freq,
-    propose_range_Q,
-    propose_range_dbGain,
-    propose_range_biquad,
-)
+from spinorama.auto_optim import optim_greedy
 from spinorama.auto_graph import graph_results as auto_graph_results
 
 
 VERSION = 0.5
-
-
-def find_best_biquad(
-    freq,
-    auto_target,
-    freq_range,
-    Q_range,
-    dbGain_range,
-    biquad_range,
-    count,
-    optim_config,
-):
-    def opt_peq(x):
-        peq = [(1.0, Biquad(int(x[0]), x[1], 48000, x[2], x[3]))]
-        return loss(freq, auto_target, peq, count, optim_config)
-
-    bounds = [
-        (biquad_range[0], biquad_range[-1]),
-        (freq_range[0], freq_range[-1]),
-        (Q_range[0], Q_range[-1]),
-        (dbGain_range[0], dbGain_range[-1]),
-    ]
-
-    logger.debug(
-        "range is [{}, {}], [{}, {}], [{}, {}], [{}, {}]".format(
-            bounds[0][0],
-            bounds[0][1],
-            bounds[1][0],
-            bounds[1][1],
-            bounds[2][0],
-            bounds[2][1],
-            bounds[3][0],
-            bounds[3][1],
-        )
-    )
-    # can use differential_evolution basinhoppin dual_annealing
-    res = {"success": False}
-    try:
-        res = opt.dual_annealing(
-            opt_peq,
-            bounds,
-            maxiter=optim_config["maxiter"],
-            # initial_temp=10000
-        )
-        logger.debug(
-            "          optim loss {:2.2f} in {} iter type {:d} at F {:.0f} Hz Q {:2.2f} dbGain {:2.2f} {}".format(
-                res.fun,
-                res.nfev,
-                int(res.x[0]),
-                res.x[1],
-                res.x[2],
-                res.x[3],
-                res.message,
-            )
-        )
-    except ValueError as ve:
-        res.success = False
-        logger.error("{} bounds {}".format(ve, bounds))
-        for i in range(0, 4):
-            if bounds[i][0] >= bounds[i][1]:
-                logger.error("on bound [{}]".format(i))
-    return res.success, int(res.x[0]), res.x[1], res.x[2], res.x[3], res.fun, res.nit
-
-
-def find_best_peak(
-    freq,
-    auto_target,
-    freq_range,
-    Q_range,
-    dbGain_range,
-    biquad_range,
-    count,
-    optim_config,
-):
-
-    biquad_type = 3
-
-    def opt_peq(x):
-        peq = [(1.0, Biquad(biquad_type, x[0], 48000, x[1], x[2]))]
-        return loss(freq, auto_target, peq, count, optim_config)
-
-    bounds = [
-        (freq_range[0], freq_range[-1]),
-        (Q_range[0], Q_range[-1]),
-        (dbGain_range[0], dbGain_range[-1]),
-    ]
-
-    logger.debug(
-        "range is [{}, {}], [{}, {}], [{}, {}]".format(
-            bounds[0][0],
-            bounds[0][1],
-            bounds[1][0],
-            bounds[1][1],
-            bounds[2][0],
-            bounds[2][1],
-        )
-    )
-    # can use differential_evolution basinhoppin dual_annealing
-    res = {"success": False}
-    try:
-        res = opt.dual_annealing(
-            opt_peq, bounds, maxiter=optim_config["maxiter"], initial_temp=10000
-        )
-        logger.debug(
-            "          optim loss {:2.2f} in {} iter type PK at F {:.0f} Hz Q {:2.2f} dbGain {:2.2f} {}".format(
-                res.fun, res.nfev, res.x[0], res.x[1], res.x[2], res.message
-            )
-        )
-        return res.success, biquad_type, res.x[0], res.x[1], res.x[2], res.fun, res.nit
-    except ValueError as ve:
-        logger.error("{} bounds {}".format(ve, bounds))
-        for i in range(0, 4):
-            if bounds[i][0] >= bounds[i][1]:
-                logger.error("on bound [{}]".format(i))
-        return False, 0, -1, -1, -1, -1, -1
-
-
-# ------------------------------------------------------------------------------
-# find next best biquad
-# ------------------------------------------------------------------------------
-
-
-def optim_preflight(freq, target, auto_target_interp, optim_config):
-    sz = len(freq)
-    nbt = len(target)
-    nbi = len(auto_target_interp)
-
-    status = True
-
-    if sz != len(target[0]):
-        logger.error("Size mismatch #freq {} != #target {}".format(sz, len(target[0])))
-        status = False
-
-    if nbt != nbi:
-        logger.error(
-            "Size mismatch #target {} != #auto_target_interp {}".format(nbt, nbi)
-        )
-        status = False
-
-    for i in range(0, len(target)):
-        if len(target[i]) != len(auto_target_interp[i]):
-            logger.error(
-                "Size mismatch #target[{}] {} != #auto_target_interp[{}] {}".format(
-                    i, len(target[i]), i, len(auto_target_interp[i])
-                )
-            )
-            status = False
-
-    return status
-
-
-def optim_compute_auto_target(freq, target, auto_target_interp, peq):
-    peq_freq = peq_build(freq, peq)
-    return [target[i] - auto_target_interp[i] + peq_freq for i in range(0, len(target))]
-
-
-def optim_greedy(
-    speaker_name, df_speaker, freq, auto_target, auto_target_interp, optim_config
-):
-
-    if optim_preflight(freq, auto_target, auto_target_interp, optim_config) is False:
-        logger.error("Preflight check failed!")
-        return None
-
-    auto_peq = []
-    current_auto_target = optim_compute_auto_target(
-        freq, auto_target, auto_target_interp, auto_peq
-    )
-    best_loss = loss(freq, auto_target, auto_peq, 0, optim_config)
-    pref_score = score_loss(df_speaker, auto_peq)
-
-    results = [(0, best_loss, -pref_score)]
-    logger.info(
-        "OPTIM {} START {} #PEQ {:d} Freq #{:d} Gain #{:d} +/-[{}, {}] Q #{} [{}, {}] Loss {:2.2f} Score {:2.2f}".format(
-            speaker_name,
-            optim_config["curve_names"],
-            optim_config["MAX_NUMBER_PEQ"],
-            optim_config["MAX_STEPS_FREQ"],
-            optim_config["MAX_STEPS_DBGAIN"],
-            optim_config["MIN_DBGAIN"],
-            optim_config["MAX_DBGAIN"],
-            optim_config["MAX_STEPS_Q"],
-            optim_config["MIN_Q"],
-            optim_config["MAX_Q"],
-            best_loss,
-            -pref_score,
-        )
-    )
-
-    for optim_iter in range(0, optim_config["MAX_NUMBER_PEQ"]):
-
-        # we are optimizing above my_freq_reg_min hz on anechoic data
-        current_auto_target = optim_compute_auto_target(
-            freq, auto_target, auto_target_interp, auto_peq
-        )
-
-        # greedy strategy: look for lowest & highest peak
-        sign, init_freq, init_freq_range = propose_range_freq(
-            freq, current_auto_target[0], optim_config
-        )
-        init_dbGain_range = propose_range_dbGain(
-            freq, current_auto_target[0], sign, init_freq, optim_config
-        )
-        init_Q_range = propose_range_Q(optim_config)
-        biquad_range = propose_range_biquad(optim_config)
-
-        (
-            state,
-            current_type,
-            current_freq,
-            current_Q,
-            current_dbGain,
-            current_loss,
-            current_nit,
-        ) = (False, -1, -1, -1, -1, -1, -1)
-        if optim_config["full_biquad_optim"] is True:
-            (
-                state,
-                current_type,
-                current_freq,
-                current_Q,
-                current_dbGain,
-                current_loss,
-                current_nit,
-            ) = find_best_biquad(
-                freq,
-                current_auto_target,
-                init_freq_range,
-                init_Q_range,
-                init_dbGain_range,
-                biquad_range,
-                optim_iter,
-                optim_config,
-            )
-        else:
-            (
-                state,
-                current_type,
-                current_freq,
-                current_Q,
-                current_dbGain,
-                current_loss,
-                current_nit,
-            ) = find_best_peak(
-                freq,
-                current_auto_target,
-                init_freq_range,
-                init_Q_range,
-                init_dbGain_range,
-                biquad_range,
-                optim_iter,
-                optim_config,
-            )
-        if state:
-            biquad = (
-                1.0,
-                Biquad(current_type, current_freq, 48000, current_Q, current_dbGain),
-            )
-            auto_peq.append(biquad)
-            best_loss = current_loss
-            pref_score = score_loss(df_speaker, auto_peq)
-            results.append((optim_iter + 1, best_loss, -pref_score))
-            logger.info(
-                "Speaker {} Iter {:2d} Optim converged loss {:2.2f} pref score {:2.2f} biquad {:2s} F:{:5.0f}Hz Q:{:2.2f} G:{:+2.2f}dB in {} iterations".format(
-                    speaker_name,
-                    optim_iter,
-                    best_loss,
-                    -pref_score,
-                    biquad[1].type2str(),
-                    current_freq,
-                    current_Q,
-                    current_dbGain,
-                    current_nit,
-                )
-            )
-        else:
-            logger.error(
-                "Speaker {} Skip failed optim for best {:2.2f} current {:2.2f}".format(
-                    speaker_name, best_loss, current_loss
-                )
-            )
-            break
-
-    # recompute auto_target with the full auto_peq
-    current_auto_target = optim_compute_auto_target(
-        freq, auto_target, auto_target_interp, auto_peq
-    )
-    final_loss = loss(freq, current_auto_target, [], 0, optim_config)
-    final_score = score_loss(df_speaker, auto_peq)
-    results.append((optim_config["MAX_NUMBER_PEQ"] + 1, best_loss, -pref_score))
-    logger.info(
-        "OPTIM END {}: best loss {:2.2f} final score {:2.2f} with {:2d} PEQs".format(
-            speaker_name, final_loss, -final_score, len(auto_peq)
-        )
-    )
-    return results, auto_peq
 
 
 @ray.remote
